topApp/requirements.txt
Django==1.8
django-cors-headers==1.1.0
django-tastypie==0.13.3
MySQL-python==1.2.5
python-dateutil==2.5.3
python-mimeparse==1.5.2
six==1.10.0
